Add new RadCaliDlg implementation and update types
Added a new source file 'RadCaliDlg - 实现 (2).cpp' containing the implementation of the CRadCaliDlg dialog and related processing logic. Updated OlpFile.hpp to use WORD* instead of short* for pixel value arrays and cast values to float in OBV construction. Modified RadCali.vcxproj to add Vcpkg configuration, enable OpenMP support, and adjust browse information settings.

+++ OlpFile.hpp
@@ -34,9 +34,10 @@ public:
     void SetSize(int sz){ m_obs.SetSize(sz); }    
     OBV* GetData(int *sz){ return m_obs.GetData(*sz); }   
     int  Append( OBV obv ){ return  m_obs.Append(obv); }
-    int  Append( double sz ,double  vz,double  as,int cc,int cr,short* cv,
-                 double sz1,double vz1,double as1,int rc,int rr,short* rv ){
-        OBV obv={ cc,cr,rc,rr,cv[0],cv[1],cv[2],cv[3],rv[0],rv[1],rv[2],rv[3],sz,vz,as,sz1,vz1,as1 };
+    int  Append( double sz ,double  vz,double  as,int cc,int cr,WORD* cv,
+                 double sz1,double vz1,double as1,int rc,int rr,WORD* rv ){
+        OBV obv={ cc,cr,rc,rr,(float)cv[0],(float)cv[1],(float)cv[2],(float)cv[3],
+            (float)rv[0],(float)rv[1],(float)rv[2],(float)rv[3],sz,vz,as,sz1,vz1,as1 };
         return Append(obv);
     };
